requirements.txt
alive-progress==1.6.1
attrs==20.3.0
coverage==5.3
cycler==0.10.0
graphviz==0.15
httpagentparser==1.9.0
iniconfig==1.1.1
kiwisolver==1.3.1
matplotlib==3.3.3
numpy==1.19.4
packaging==20.4
Pillow==8.0.1
pluggy==0.13.1
pprintpp==0.4.0
py==1.9.0
pycodestyle==2.6.0
pycountry==20.7.3
pycountry-convert==0.7.2
pyparsing==2.4.7
pytest==6.1.2
pytest-cov==2.10.1
pytest-mock==3.3.1
python-dateutil==2.8.1
PyYAML==5.3.1
repoze.lru==0.7
six==1.15.0
toml==0.10.2
ua-parser==0.10.0
user-agents==2.2.0
